Completely revamp a bunch of stuff
Auth is now moved out of the index file and into the App.

It'll probably be moved from within there to somewhere else too,
just not sure where.

Also I am renaming BasicInfo to AuthInfo to make it more explicit,
and now User does not reference any of those info things directly,
and instead just contains the info itself. I think that makes more sense.

Oh, and Home builds the list of worlds from the user to display.

And probably more because I waited way too long to commit these files
and it's now a bit of a mess.