Allow going to world page without logging in
If I do a redirect, then the user will actually end up starting
on the home page no matter what.

So instead I'm just letting them load the page and it just shows
nothing.

This is mainly an issue because at first load the user won't load in,
so on a reload the app would redirect them back to home. Now
instead it will just quickly sit on that page and if they don't do
anything it will all load.