Upgrade parcel to v2
This is being upgraded because the v1 version of
bundler isn't getting any new versions being released
and I want to get rid of the dependency alerts.

Once I did this, the paths that I had defined stopped
working, and since I wasn't in love with them as it was
I decided to just remove them.